requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
numpy==1.25.2
requests==2.31.0
gunicorn==21.2.0
scikit-learn==1.3.0
Pillow==10.0.1
tensorflow-cpu==2.13.0